Enum soundchange-english::Phoneme

pub enum Phoneme {
    P,
    B,
    T,
    D,
    G,
    K,
    M,
    N,
    NG,
    F,
    V,
    TH,
    S,
    Z,
    SH,
    CH,
    J,
    R,
    L,
    H,
    Y,
    W,
    _C,
    _Q,
    _X,
    AY,
    A,
    EE,
    E,
    AI,
    I,
    OH,
    O,
    EW,
    U,
    OO,
    AW,
    UU,
    UH,
    _A,
    _E,
    _I,
    _O,
}

A phoneme for English.

While variants are named after their common respelling (slightly adjusted to fit in the ASCII and to be unique) they do not directly represent an unique sound (say, voiceless bilabial stop /p/). A phoneme rather represents an (conceptual) equivalence class for related sounds, so that it is a minimal unit that can give the difference in meaning. Different dialects of English commonly have different sounds for the same phoneme.

Variants

P

B

T

D

G

K

M

N

NG

F

V

TH

S

Z

SH

CH

J

R

L

H

Y

W

_C

_Q

_X

AY

A

EE

E

AI

I

OH

O

EW

U

OO

AW

UU

UH

_A

_E

_I

_O

Methods

impl Phoneme

fn to_char(&self) -> char

fn from_char(c: char) -> Option<Phoneme>

fn is_vowel(&self) -> bool

fn is_consonant(&self) -> bool

fn is_untranslated(&self) -> bool

fn is_short_vowel(&self) -> bool

fn is_long_vowel(&self) -> bool

fn to_short_vowel(&self) -> Phoneme

fn to_long_vowel(&self) -> Phoneme